src/main/openobserve/api_tests/requirements.txt
# OpenObserve API test dependencies
# UI tests in src/main/openobserve/ui-tests use npm dependencies.

pytest==8.3.5
requests==2.32.3
python-dotenv==1.0.1
allure-pytest==2.15.0
pytest-html==4.1.1
pytest-timeout==2.4.0
pytest-metadata==3.1.1

